requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
sentence-transformers
python-dotenv
faiss-cpu